refactor: Change PackageExtractor to BasePackager; refactor folder location.
