+++ filejacket/pipelines/packager.py
@@ -23,7 +23,7 @@ Should there be a need for contact the electronic mail
 from __future__ import annotations
 
 from datetime import datetime
-from io import IOBase, BytesIO
+from io import BytesIO
 from sys import getsizeof
 from tarfile import TarFile, TarError
 from typing import Any, TYPE_CHECKING, Type, IO
@@ -31,20 +31,18 @@ from zipfile import BadZipFile, ZipFile
 
 from rarfile import BadRarFile, RarFile, NotRarFile
 
-from ..base import BaseExtractor
-from ..hasher import CRC32Hasher
-from ...adapters.pipeline import PipelineSequential
-from ...exception import ValidationError
-from ...utils import LazyImportClass
+from .base import BasePackager
+from .hasher import CRC32Hasher
+from ..adapters.pipeline import PipelineSequential
+from ..utils import LazyImportClass
 
 if TYPE_CHECKING:
-    from ...file import BaseFile
-    from ...engines.storage import StorageEngine
+    from ..file import BaseFile
+    from ..engines.storage import StorageEngine
     from psd_tools import PSDImage
     from py7zr import SevenZipFile, FileInfo
 
 __all__ = [
-    "PackageExtractor",
     "PSDLayersFromPackageExtractor",
     "SevenZipCompressedFilesFromPackageExtractor",
     "RarCompressedFilesFromPackageExtractor",
@@ -53,229 +51,20 @@ __all__ = [
 ]
 
 
-class PackageExtractor(BaseExtractor):
-    """
-    Extractor class with focus to processing information from file's content.
-    This class was created to allow parsing and extraction of data designated to
-    internal files.
-    """
-
-    extensions: set[str]
-    extensions = None
-    """
-    Attribute to store allowed extensions for use in `validator`.
-    This attribute should be override in children classes.
-    """
-    compressor_class: Type[type]
-    compressor_class = None
-    """
-    Attribute to store the current class of compressor for use in `content_buffer` and `decompress` methods.
-    This attribute should be override in children classes.
-    """
-    stopper: bool = True
-    """
-    Variable that define if this class used as processor should stop the pipeline.
-    """
-
-    class ContentBuffer(IOBase):
-        """
-        Class to allow consumption of buffer in a lazy way.
-        This class should be override in children of PackageExtractor to
-        implementation of method read().
-        """
-
-        source_file_object: BaseFile
-        source_file_object = None
-        """
-        Attribute to store the related file object that has the buffer for the compressed content.
-        """
-        compressor: Type[type]
-        compressor = None
-        """
-        Attribute to store the class of the compressor able to uncompress the content.  
-        """
-        compressed_object: Any
-        compressed_object = None
-        """
-        Attribute to store the instance of the compressor.
-        """
-        filename: str
-        filename = None
-        """
-        Attribute to store the name of file that should be extract for this content.
-        """
-        mode: str
-        mode = None
-        """
-        Attribute to store the mode of read for the uncompressed content. 
-        """
-
-        reference_class: Type[PackageExtractor]
-        reference_class = None
-        """
-        Attribute to allow serialization of this  class as local class.
-        """
-
-        buffer: Any
-        """
-        Attribute to store the current initialized buffer.
-        """
-
-        def __init__(
-            self: PackageExtractor.ContentBuffer,
-            source_file_object: BaseFile,
-            compressor_class: Type[type],
-            internal_file_filename: str,
-            mode: str,
-            reference: Type[PackageExtractor],
-        ) -> None:
-            """
-            Method to initiate the object saving the data required to allow decompressing and reading content
-            for specific file.
-            """
-            self.source_file_object = source_file_object
-            self.compressor = compressor_class
-            self.filename = internal_file_filename
-            self.mode = mode
-            self.reference = reference
-
-        def read(
-            self: PackageExtractor.ContentBuffer, *args: Any, **kwargs: Any
-        ) -> str | bytes:
-            """
-            Method to read the content of the object initiating the buffer if not exists.
-            """
-            if not hasattr(self, "buffer"):
-                # Instantiate the buffer of inner content
-                self.mount_buffer()
-
-            return self.buffer.read(*args, **kwargs)
-
-        def mount_buffer(self: PackageExtractor.ContentBuffer) -> None:
-            """
-            Method to initiate the buffer object if not exists.
-            This method should be overwritten in child class.
-            """
-            raise NotImplementedError(
-                f"Method mount_buffer of PackageExtractor.ContentBuffer should be override in child class "
-                f"{self.__class__.__name__}."
-            )
-
-        def seek(self, *args: Any, **kwargs: Any) -> int:
-            """
-            Method to seek the content in the buffer.
-            Buffer must exist for this method to work, else no action will be taken.
-            """
-            if not hasattr(self, "buffer"):
-                # Initiate the buffer
-                # It will begin extraction of file to have access to its buffer.
-                self.mount_buffer()
-
-            return self.buffer.seek(*args, **kwargs)
-
-        def seekable(self) -> bool:
-            """
-            Method to verify if buffer is seekable.
-            Buffer must exist for this method to work, else no action will be taken.
-            
-            For better performance this method should be override in child class to avoid using buffer, as it extract the content
-            in memory.
-            """
-            if not hasattr(self, "buffer"):
-                # Initiate the buffer
-                # It will begin extraction of file to have access to its buffer.
-                self.mount_buffer()
-
-            return self.buffer.seekable()
-
-        def close(self) -> None:
-            """
-            Method to close the buffer.
-            Buffer must exist for this method to work, else no action will be taken.
-            """
-            if not hasattr(self, "buffer"):
-                return
-
-            self.buffer.close()
-            delattr(self, "buffer")
-
-    @classmethod
-    def validate(cls, file_object: BaseFile) -> None:
-        """
-        Method to validate if content can be extract to given extension.
-        """
-        if cls.extensions is None:
-            raise NotImplementedError(
-                f"The attribute `extensions` is not overwritten in child class {cls.__name__}"
-            )
-
-        if cls.compressor_class is None:
-            raise NotImplementedError(
-                f"The attribute `compressor_class` is not overwritten in child class {cls.__name__}"
-            )
-
-        # The ValidationError should be captured in children classes else it will not register as an error and
-        # the pipeline will break.
-        if file_object.extension not in cls.extensions:
-            raise ValidationError(
-                f"Extension `{file_object.extension}` not allowed in validate for class {cls.__name__}"
-            )
-
-    @classmethod
-    def decompress(cls, file_object: BaseFile, overrider: bool, **kwargs: Any) -> bool:
-        """
-        Method to uncompress the content from a file_object.
-        This method must be override in child class.
-        """
-        raise NotImplementedError(
-            "Method extract_content must be overwritten on child class."
-        )
-
-    @classmethod
-    def content_buffer(
-        cls, file_object: BaseFile, internal_file_name: str, mode: str = "rb"
-    ) -> ContentBuffer:
-        """
-        Method to create a buffer pointing to the uncompressed content.
-        This method must work lazily, extracting the content only when the buffer is read.
-        This method must be override in child class.
-        """
-        raise NotImplementedError(
-            "Method content_buffer must be overwritten on child class."
-        )
-
-    @classmethod
-    def process(cls, **kwargs: Any) -> bool:
-        """
-        Method used to run this class on Processor`s Pipeline for Extracting info from Data.
-        This process method is created exclusively to pipeline for objects inherent from BaseFile.
-
-        The processor for package extraction override the method `BaseExtractor.process` in order to validate
-        the extension before processing the `object_to_process`.
-        """
-        try:
-            object_to_process: BaseFile = kwargs["object_to_process"]
-            cls.validate(file_object=object_to_process)
-        except (ValidationError, KeyError):
-            return False
-
-        return super().process(**kwargs)
-
-
-class MastrokaFilesFromPackageExtractor(PackageExtractor):
+class MastrokaFilesFromPackageExtractor(BasePackager):
     """
     Class to extract internal files from mka, mkv files.
     """
 
 
-class PDFPagesFromPackageExtractor(PackageExtractor):
+class PDFPagesFromPackageExtractor(BasePackager):
     """
     Class to extract internal files from PDF files.
     """
 
 
 
-class PSDLayersFromPackageExtractor(PackageExtractor):
+class PSDLayersFromPackageExtractor(BasePackager):
     """
     Class to extract internal files from PSD files.
     """
@@ -292,18 +81,18 @@ class PSDLayersFromPackageExtractor(PackageExtractor):
     @classmethod
     def content_buffer(
         cls, file_object: BaseFile, internal_file_name: str, mode: str = "rb"
-    ) -> PackageExtractor.ContentBuffer:
+    ) -> BasePackager.ContentBuffer:
         """
         Method to create a buffer pointing to the uncompressed content.
         This method must work lazily, extracting the content only when the buffer is read.
         """
 
-        class PSDContentBuffer(PackageExtractor.ContentBuffer):
+        class PSDContentBuffer(BasePackager.ContentBuffer):
             """
             Class to allow consumption of buffer in a lazy way.
             """
 
-            def mount_buffer(self: PackageExtractor.ContentBuffer) -> None:
+            def mount_buffer(self: BasePackager.ContentBuffer) -> None:
                 """
                 Method to initiate the buffer object if not exists.
                 """
@@ -447,7 +236,7 @@ class PSDLayersFromPackageExtractor(PackageExtractor):
         return True
 
 
-class TarCompressedFilesFromPackageExtractor(PackageExtractor):
+class TarCompressedFilesFromPackageExtractor(BasePackager):
     """
     Class to extract internal files from tar.gz or tar.bz files.
     As gzip and bzip compression don`t provide a manifest, as it is not an archive format just a compression algorithm,
@@ -467,18 +256,18 @@ class TarCompressedFilesFromPackageExtractor(PackageExtractor):
     @classmethod
     def content_buffer(
         cls, file_object: BaseFile, internal_file_name: str, mode: str = "rb"
-    ) -> PackageExtractor.ContentBuffer:
+    ) -> BasePackager.ContentBuffer:
         """
         Method to create a buffer pointing to the uncompressed content.
         This method must work lazily, extracting the content only when the buffer is read.
         """
 
-        class TarContentBuffer(PackageExtractor.ContentBuffer):
+        class TarContentBuffer(BasePackager.ContentBuffer):
             """
             Class to allow consumption of buffer in a lazy way.
             """
 
-            def mount_compressed_object(self: PackageExtractor.ContentBuffer) -> None:
+            def mount_compressed_object(self: BasePackager.ContentBuffer) -> None:
                 """
                 Method to initialize the compressed file from the upstream package buffer.
                 """
@@ -487,7 +276,7 @@ class TarCompressedFilesFromPackageExtractor(PackageExtractor):
                     fileobj=self.source_file_object.content_as_buffer
                 )
                 
-            def mount_buffer(self: PackageExtractor.ContentBuffer) -> None:
+            def mount_buffer(self: BasePackager.ContentBuffer) -> None:
                 """
                 Method to initiate the buffer object if not exists.
                 """
@@ -670,7 +459,7 @@ class TarCompressedFilesFromPackageExtractor(PackageExtractor):
         return True
 
 
-class ZipCompressedFilesFromPackageExtractor(PackageExtractor):
+class ZipCompressedFilesFromPackageExtractor(BasePackager):
     """
     Class to extract internal files from zip and cbz files.
     """
@@ -696,7 +485,7 @@ class ZipCompressedFilesFromPackageExtractor(PackageExtractor):
             Class to allow consumption of buffer in a lazy way.
             """
 
-            def mount_compressed_object(self: PackageExtractor.ContentBuffer) -> None:
+            def mount_compressed_object(self: BasePackager.ContentBuffer) -> None:
                 """
                 Method to initialize the compressed file from the upstream package buffer.
                 """
@@ -705,7 +494,7 @@ class ZipCompressedFilesFromPackageExtractor(PackageExtractor):
                     file=self.source_file_object.content_as_buffer
                 )
                 
-            def mount_buffer(self: PackageExtractor.ContentBuffer) -> None:
+            def mount_buffer(self: BasePackager.ContentBuffer) -> None:
                 """
                 Method to initiate the buffer object if not exists.
                 """
@@ -868,7 +657,7 @@ class ZipCompressedFilesFromPackageExtractor(PackageExtractor):
         return True
 
 
-class RarCompressedFilesFromPackageExtractor(PackageExtractor):
+class RarCompressedFilesFromPackageExtractor(BasePackager):
     """
     Class to extract internal files from rar files.
     """
@@ -885,7 +674,7 @@ class RarCompressedFilesFromPackageExtractor(PackageExtractor):
     @classmethod
     def content_buffer(
         cls, file_object: BaseFile, internal_file_name: str, mode: str = "rb"
-    ) -> PackageExtractor.ContentBuffer:
+    ) -> BasePackager.ContentBuffer:
         """
         Method to create a buffer pointing to the uncompressed content.
         This method must work lazily, extracting the content only when the buffer is read.
@@ -896,7 +685,7 @@ class RarCompressedFilesFromPackageExtractor(PackageExtractor):
             Class to allow consumption of buffer in a lazy way.
             """
 
-            def mount_compressed_object(self: PackageExtractor.ContentBuffer) -> None:
+            def mount_compressed_object(self: BasePackager.ContentBuffer) -> None:
                 """
                 Method to initialize the compressed file from the upstream package buffer.
                 """
@@ -905,7 +694,7 @@ class RarCompressedFilesFromPackageExtractor(PackageExtractor):
                     file=self.source_file_object.content_as_buffer
                 )
                 
-            def mount_buffer(self: PackageExtractor.ContentBuffer) -> None:
+            def mount_buffer(self: BasePackager.ContentBuffer) -> None:
                 """
                 Method to initiate the buffer object if not exists.
                 """
@@ -1068,7 +857,7 @@ class RarCompressedFilesFromPackageExtractor(PackageExtractor):
         return True
 
 
-class SevenZipCompressedFilesFromPackageExtractor(PackageExtractor):
+class SevenZipCompressedFilesFromPackageExtractor(BasePackager):
     """
     Class to extract internal files from 7z files.
     """
@@ -1087,18 +876,18 @@ class SevenZipCompressedFilesFromPackageExtractor(PackageExtractor):
     @classmethod
     def content_buffer(
         cls, file_object: BaseFile, internal_file_name: str, mode: str = "rb"
-    ) -> PackageExtractor.ContentBuffer:
+    ) -> BasePackager.ContentBuffer:
         """
         Method to create a buffer pointing to the uncompressed content.
         This method must work lazily, extracting the content only when the buffer is read.
         """
 
-        class SevenZipContentBuffer(PackageExtractor.ContentBuffer):
+        class SevenZipContentBuffer(BasePackager.ContentBuffer):
             """
             Class to allow consumption of buffer in a lazy way.
             """
 
-            def mount_compressed_object(self: PackageExtractor.ContentBuffer) -> None:
+            def mount_compressed_object(self: BasePackager.ContentBuffer) -> None:
                 """
                 Method to initialize the compressed file from the upstream package buffer.
                 """
@@ -1107,7 +896,7 @@ class SevenZipCompressedFilesFromPackageExtractor(PackageExtractor):
                     file=self.source_file_object.content_as_buffer
                 )  # type: ignore
             
-            def mount_buffer(self: PackageExtractor.ContentBuffer) -> None:
+            def mount_buffer(self: BasePackager.ContentBuffer) -> None:
                 """
                 Method to initiate the buffer object if not exists.
                 """
