+++ filejacket/pipelines/render/static.py
@@ -26,7 +26,7 @@ from io import BytesIO, StringIO
 from typing import Any, TYPE_CHECKING, Type
 
 from ..base import BaseRender
-from ... import PipelineSequential
+from ...adapters.pipeline import PipelineSequential
 from ...exception import RenderError
 
 if TYPE_CHECKING:
